delete order

+++ controllers/orderControllers.js
@@ -74,12 +74,18 @@ class orderController{
     }
     static deleteOrder=async (req,res)=>{
         const {id}=req.body
-        try {
-            await orderModel.findByIdAndDelete(id)
-            res.send({"status":"success","message":"delete order successfully..."})
-        } catch (error) {
-            res.send({"status":"failed","message":"Unable to delete order...."})
+        const order=await orderModel.findById(id)
+        if(!order){
+            res.send({"status":"failed","message":"Order id is incorrect...."})
         }
+        else{
+            try {
+                await orderModel.findByIdAndDelete(id)
+                res.send({"status":"success","message":"delete order successfully..."})
+            } catch (error) {
+                res.send({"status":"failed","message":"Unable to delete order...."})
+            }
+        }    
     }
     static getAllOrder=async (req,res)=>{
         try {
